Fix truncated comparison with correct types

+++ crates/leta-daemon/src/handlers/grep.rs
@@ -173,7 +173,7 @@ pub async fn handle_grep(ctx: &HandlerContext, params: GrepParams) -> Result<Gre
         None
     };
 
-    let truncated = filtered.len() as u32 >= limit;
+    let truncated = filtered.len() >= limit;
 
     Ok(GrepResult {
         symbols: filtered,
